refactor(llm): reorganize runtime domains
