+++ packages/llm/AGENTS.md
@@ -9,27 +9,31 @@ src/
 ├── index.ts          # Narrow public API: Auth, Provider, ModelsDev, errors, run, RunInput, TokenTracker
 ├── run.ts            # run() — model-required top-level entry: messages+tools → Run.Outcome via Sink
 ├── error.ts          # Re-exports NamedError classes from protocol
+├── message/
+│   └── index.ts      # toModelMessages() — Message.WithParts[] → AI SDK messages
+├── processor/
+│   ├── index.ts      # Processor.create() orchestrator — bounded retry loop + stream driving
+│   ├── stream-events.ts # Stream event dispatch for text/reasoning/tool/step events
+│   ├── sink-projection.ts # Sink projection to Bus telemetry + noop sink
+│   ├── step-accounting.ts # step-start / step-finish token accounting
+│   ├── tool-projection.ts # tool-call/tool-result projection, optional execution, interruption cleanup
+│   └── contracts.ts  # Internal Processor implementation contracts
+├── retry/
+│   └── index.ts      # Retry.delay / sleep / isRetryable — exponential backoff + retry-after
 ├── session/
-│   ├── processor.ts  # Processor.create() namespace barrel/orchestrator — bounded retry loop + stream driving
-│   ├── processor-events.ts # Stream event dispatch for text/reasoning/tool/step events
-│   ├── processor-sink.ts # Sink projection to Bus telemetry + noop sink
-│   ├── processor-steps.ts # step-start / step-finish token and cost accounting
-│   ├── processor-tools.ts # tool-call/tool-result projection, optional execution, interruption cleanup
-│   ├── processor-types.ts # Internal Processor implementation contracts
-│   ├── index.ts      # Deep import surface for protocol Message/Tool plus session helpers
-│   ├── convert.ts    # toModelMessages() — Message.WithParts[] → AI SDK messages
-│   └── retry.ts      # Retry.delay / sleep / isRetryable — exponential backoff + retry-after
+│   └── index.ts      # Compatibility facade only; do not add new session responsibilities here
 ├── auth/
 │   ├── storage.ts    # Auth namespace: get / set / remove / all (credential storage)
 ├── provider/
 │   ├── index.ts      # Provider + ModelsDev public namespaces; internal provider helpers stay deep
-│   └── provider.ts   # getSDK() + getLanguage() — maps Provider.Model to @ai-sdk/* instance
-├── transform/
-│   └── index.ts      # ProviderTransform — message normalization, caching, per-provider variants
+│   ├── sdk.ts        # getSDK() + getLanguage() — maps Provider.Model to @ai-sdk/* instance
+│   ├── transform.ts  # ProviderTransform — message normalization, caching, per-provider variants
+│   └── proxy-models.ts # Proxy model catalog fetch/enrichment
 ├── token/
 │   └── index.ts      # TokenTracker.extractUsage
 ├── model/
-│   └── index.ts      # ModelsDev.get / refresh — fetches models.dev catalog
+│   ├── index.ts      # ModelsDev.get / refresh — fetches models.dev catalog
+│   └── models-snapshot.json # Bundled trusted catalog snapshot
 └── util/
     └── lazy.ts       # Lazy initialization helper
 ```
@@ -41,7 +45,7 @@ src/
 - **Provider.Model**: Zod schema with capabilities, cost, limits, status. Built from `models.dev` data via `Provider.fromModelsDevModel()`. `Provider.listModels()` / `listProviders()` / `getProviderInfo()` surface catalog lookups.
 - **Auth.Info** (discriminated union): `{ type: "api", key }` | `{ type: "proxy", baseURL, apiKey? }`. Stored via `Auth.set(providerId, info)` and read by `getSDK()` before each call.
 - **SDK wiring**: `getSDK(model, auth)` resolves to Anthropic / OpenAI. Custom OpenAI-compatible endpoints use `@ai-sdk/openai` with `baseURL` / `name`, keeping returned language models on the same AI SDK provider type version. SDK and `LanguageModel` instances are cached per `providerID:npm:modelID:auth` key. Provider-specific behavior belongs in `provider/`, `auth/`, or `transform/`, not in call sites.
-- **Provider transforms** (`transform/index.ts`): `normalizeMessages()` filters empty blocks, sanitizes tool-call IDs, applies Anthropic ephemeral caching to the last two user/assistant messages. `variants(model)` exposes per-provider thinking / reasoning presets; `resolveVariant(model, variant?)` picks one. This is an internal/deep import surface, not a root export.
+- **Provider transforms** (`provider/transform.ts`): `normalizeMessages()` filters empty blocks, sanitizes tool-call IDs, applies Anthropic ephemeral caching to the last two user/assistant messages. `variants(model)` exposes per-provider thinking / reasoning presets; `resolveVariant(model, variant?)` picks one. This is an internal/deep import surface, not a root export.
 - **Processor**: Created via `Processor.create({ assistantMessage, sessionID, model, abort, sink, onToolCall, createStream, maxRetryAttempts? })`. `process()` returns `"stop" | "continue" | "compact"`. Accumulates `TextPart` / `ReasoningPart` / `ToolPart` and publishes through `Sink`. Processor owns `tool-call` and `tool-result` stream projection; `run()`'s AI SDK `execute` callback must not directly emit tool sink events.
 - **Retry**: `Retry.delay(attempt, error?)` computes backoff respecting `retry-after` / `retry-after-ms` headers. `Retry.isRetryable(error)` checks `APIError.isRetryable`. Processor retrying is finite by default; do not use unbounded retry loops or publish `Number.MAX_SAFE_INTEGER` as a retry cap.
 - **TokenTracker**: Extracts token usage from AI SDK/provider responses. Runtime accounting stores token counts on the assistant message, keyed by that message's provider/model identity. The llm package does not calculate dollar cost.
