+++ packages/llm/src/model/models-snapshot.json
