+++ packages/llm/src/processor/contracts.ts
