+++ packages/llm/src/processor/sink-projection.ts
@@ -1,6 +1,6 @@
 import { Operational, type Run, type Sink, type Tool } from "@openomni/protocol";
 import { Bus } from "@openomni/session";
-import { generateId } from "./processor-types.js";
+import { generateId } from "./contracts.js";
 
 export interface ProjectedSink extends Sink {
   flush(): Promise<void>;
