+++ packages/llm/src/processor/step-accounting.ts
@@ -1,7 +1,7 @@
 import type { Message } from "@openomni/protocol";
 import { TokenTracker } from "../token";
 import type { Provider } from "../provider";
-import { generateId, type StreamEvent } from "./processor-types.js";
+import { generateId, type StreamEvent } from "./contracts.js";
 
 export type StepEventContext = {
   readonly sessionID: string;
