+++ packages/llm/test/processor/tool-result.test.ts
@@ -1,8 +1,7 @@
 import { afterEach, describe, expect, test } from "bun:test";
-import type { Run, Sink, Tool } from "@openomni/protocol";
+import type { Message, Run, Sink, Tool } from "@openomni/protocol";
 import { Bus } from "@openomni/session";
-import type { Message } from "../../src/session";
-import { Processor } from "../../src/session/processor";
+import { Processor } from "../../src/processor";
 import type { Provider } from "../../src/provider";
 
 function assistantMessage(): Message.AssistantMessage {
